requirements.txt
bs4
lxml
cssselect
Pillow
requests
